完成了volunteer端口的所有功能

+++ src/main/java/com/neu/dreambuilder/service/volunteer/VolunteerMissionService.java
@@ -18,5 +18,5 @@ public interface VolunteerMissionService {
 
     void putScore(Long missionId, Long kidId, int score);
 
-    Map<String,Integer> getMissionTatal();
+    Map<String,Long> getMissionTatal();
 }
